requirements.txt
apscheduler>=3.11.1
authlib>=1.6.5
flask>=3.1.2
flask-cors>=6.0.1
flask-jwt-extended>=4.7.1
flask-limiter>=4.0.0
flask-mail>=0.10.0
flask-migrate>=4.1.0
flask-sqlalchemy>=3.1.1
gunicorn>=23.0.0
marshmallow>=4.1.0
psycopg2-binary>=2.9.11
python-dotenv>=1.2.1
python-jose>=3.5.0
requests>=2.32.5
requests-oauthlib>=2.0.0
werkzeug>=3.1.3
cryptography>=41.0.0